moved exceptions into proper folder, fixed file paths for test
