+++ code/src/Core/Core.csproj
@@ -169,7 +169,7 @@
     <Compile Include="PostActions\Catalog\Merge\FailedMergePostAction.cs" />
     <Compile Include="Fs.cs" />
     <Compile Include="Gen\TempGenerationResult.cs" />
-    <Compile Include="Locations\InvalidSignatureException.cs" />
+    <Compile Include="Packaging\InvalidSignatureException.cs" />
     <Compile Include="Locations\LocalTemplatesSource.cs" />
     <Compile Include="Composition\CompositionQuery.cs" />
     <Compile Include="Locations\SyncStatusEventArgs.cs" />
@@ -276,7 +276,7 @@
     <Compile Include="Locations\TemplatesContent.cs" />
     <Compile Include="Locations\TemplatesSource.cs" />
     <Compile Include="Packaging\Templatex.cs" />
-    <Compile Include="Locations\SignCertNotFoundException.cs" />
+    <Compile Include="Packaging\SignCertNotFoundException.cs" />
     <Compile Include="MetadataInfo.cs" />
     <Compile Include="Mvvm\ObservableBase.cs" />
     <Compile Include="Mvvm\RelayCommand.cs" />
